fix(ci+security): address build failures and CodeQL alerts
Build fixes:
- ruby/csv_parser: add lib/coding_adventures_csv_parser.rb top-level entry
  file. `require "coding_adventures_csv_parser"` searches for a file by that
  name in the gem's lib/ dir. Without it, Ruby raises LoadError when
  sql_csv_source does `require "coding_adventures_csv_parser"`. Pattern matches
  all other gems (e.g. coding_adventures_sql_execution_engine.rb exists in
  sql_execution_engine/lib/).

- python/csv-parser: replace the fragile sys.modules runtime check with a no-op
  + comment. pytest-cov and coverage.py import stdlib csv internally before the
  test runs, making sys.modules["csv"] point to the stdlib module regardless of
  whether our code imports it. The authoritative check (reading the source file
  directly) is still present in test_parser_source_has_no_csv_import.

- typescript/sql-csv-source: generate package-lock.json. The BUILD file uses
  `npm ci` which requires a lock file to be present. Generated by running
  `npm install` locally.

Security fix (CodeQL #80, #81 — high severity, go/incorrect-integer-conversion):
- go/sql-execution-engine executor.go lines 1108/1116: converting int64
  limitVal/offsetVal to int via int(x) is unsafe on 32-bit platforms where int
  is 32-bit (CWE-190 integer overflow). Fixed by casting len(rows) to int64
  and comparing int64 values directly, which is always safe since slice length
  is bounded by addressable memory and always fits in int64.

+++ code/packages/go/sql-execution-engine/executor.go
@@ -1104,8 +1104,11 @@ func executeLimit(node *parser.ASTNode, result *QueryResult) *QueryResult {
 	rows := result.Rows
 
 	// Apply OFFSET.
+	// Compare against int64(len(rows)) to avoid narrowing int64→int on 32-bit
+	// platforms (CWE-190). len() always fits in int64 since slice length is
+	// bounded by addressable memory.
 	if offsetVal > 0 {
-		if int(offsetVal) >= len(rows) {
+		if offsetVal >= int64(len(rows)) {
 			rows = nil
 		} else {
 			rows = rows[offsetVal:]
@@ -1113,7 +1116,8 @@ func executeLimit(node *parser.ASTNode, result *QueryResult) *QueryResult {
 	}
 
 	// Apply LIMIT.
-	if limitVal >= 0 && int(limitVal) < len(rows) {
+	// Same int64 comparison to avoid narrowing conversion (CWE-190).
+	if limitVal >= 0 && limitVal < int64(len(rows)) {
 		rows = rows[:limitVal]
 	}
